+++ include/nstd/gl/state.h
@@ -40,16 +40,17 @@ typedef struct
 } NSTDGLStateDescriptor;
 
 /// Creates a new GL state.
+/// NOTE: `surface` and `device_handle` are freed once the state is freed.
 /// Parameters:
-///     `const NSTDGLInstance instance` - An instance of `wgpu`.
 ///     `const NSTDWindow window` - The window in which the GL state will live in.
 ///     `NSTDGLSurface *const surface` - The surface that the state will use.
+///     `NSTDGLDeviceHandle *const device_handle` - The device handle to create the device with.
 ///     `const NSTDGLStateDescriptor descriptor` - Configures the state.
 /// Returns: `NSTDGLState state` - The new GL state.
 NSTDAPI NSTDGLState nstd_gl_state_new(
-    const NSTDGLInstance instance,
     const NSTDWindow window,
     NSTDGLSurface *const surface,
+    NSTDGLDeviceHandle *const device_handle,
     const NSTDGLStateDescriptor descriptor);
 
 /// Pushes the current frame to the display.